refactor: cau -> univ로 폴더명 수정
